Added files to tracker
All files in Client-HL\Assets\OBJ-IO\Plugins\Mesh\OBJ\ were currently being ignored, but were necessary for the project to function. None of the model files were being tracked either. In this commit these necessary files were added to the tracker, so the project should work better when it is downloaded to new systems.