Added 6Dof functionality
This update incluides the 6DoF functionality. The main files added in this update is PlayerInput, FixedJoystick and CameraHandler. The Input settings under the project settings have been edited a bit to include the ZAxis and Mouse ZAxis. Two new UI objects RotationalJoystick and DirectionalJoystick has been added for the 6Dof controls.

This branch also updated the mobile app to 2019.4.14f4